test: extract IssueExportTestSupport (3 actor mocks) from TestSupport (#557, #434)
Seventh slice of #434 (TestSupport.swift decomposition). Extracts the 3
issue-extraction/export actor mocks — MockIssueExtractionService (16 lines),
MockExportService (264 lines), StubExportReceiptStore (77 lines incl. doc) —
into a new Tests/BugNarratorTests/IssueExportTestSupport.swift.

Byte-identical move; SHAs verified against origin/main:
- Block A: f58fa9598f51f1af8103e786a60bfc1739a710429e17c69b370104dc02ce8e47
- Block B: 70f604126632b9fcd419ee4902b51250547e74b855480cd58c5ecd90d5dfde5f
- Block C: eb4225eadc3f7cb032b095ad13533e293d5a8c4f517d1caeb142a488a5318dbc

Non-adjacent extraction protocol (per #556): reverse-order sed deletion of the
3 blocks so earlier line numbers stay valid.

Actor isolation preserved (still `actor`, not `class`) — critical concurrency
invariant flagged by codex spec-review.

Imports: `import Foundation` + `@testable import BugNarrator`. Codex spec-review
caught that `import XCTest` would be dead (no XCTest symbols in actor bodies).

TestSupport.swift now retains only the 3 free-standing helper funcs
(makeSampleTranscriptSession, waitUntil, requestBodyData) — those are absorbed
by #570 (TestFixtures extraction + shell deletion).

Tests: 667 (unchanged).
